Products data types fixed

+++ src/mag-app.Domain/Entities/Products/Product.cs
@@ -10,9 +10,9 @@ namespace mag_app.Domain.Entities.Products
     public class Product : Auditable
     {
         public string ProdutName { get; set; } = string.Empty;
-        public double Price { get; set; }
-        public double PurchasedPrice { get; set; }
-        public decimal  Quantity { get; set; }
+        public decimal Price { get; set; }
+        public decimal PurchasedPrice { get; set; }
+        public int  Quantity { get; set; }
         public string Description { get; set; } = string.Empty;
         public long SubCategoryId { get; set; }
         public virtual SubCategory SubCategory { get; set; } = default!;
